+++ evaluator/evaluate_advisories_test.go
@@ -69,21 +69,21 @@ func TestGetStoredAdvisoriesMap(t *testing.T) {
 }
 
 func TestGetNewAndUnpatchedAdvisories(t *testing.T) {
-	stored := database.CreateStoredAdvisories(map[int]*time.Time{1: &testDate, 2: nil, 3: nil})
+	stored := database.CreateStoredAdvisories(map[int64]*time.Time{1: &testDate, 2: nil, 3: nil})
 	reported := database.CreateReportedAdvisories("ER-1", "ER-3", "ER-4")
 	news, unpatched := getNewAndUnpatchedAdvisories(reported, stored)
 	assert.Equal(t, 1, len(news))
 	assert.Equal(t, "ER-4", news[0])
 	assert.Equal(t, 1, len(unpatched))
-	assert.Equal(t, 1, unpatched[0])
+	assert.Equal(t, int64(1), unpatched[0])
 }
 
 func TestGetPatchedAdvisories(t *testing.T) {
-	stored := database.CreateStoredAdvisories(map[int]*time.Time{1: &testDate, 2: nil, 3: nil})
+	stored := database.CreateStoredAdvisories(map[int64]*time.Time{1: &testDate, 2: nil, 3: nil})
 	reported := database.CreateReportedAdvisories("ER-3", "ER-4")
 	patched := getPatchedAdvisories(reported, stored)
 	assert.Equal(t, 1, len(patched))
-	assert.Equal(t, 2, patched[0])
+	assert.Equal(t, int64(2), patched[0])
 }
 
 func TestUpdatePatchedSystemAdvisories(t *testing.T) {
@@ -91,12 +91,12 @@ func TestUpdatePatchedSystemAdvisories(t *testing.T) {
 	core.SetupTestEnvironment()
 
 	system := models.SystemPlatform{ID: 12, RhAccountID: 3}
-	advisoryIDs := []int{2, 3, 4}
+	advisoryIDs := []int64{2, 3, 4}
 	database.CreateSystemAdvisories(t, system.RhAccountID, system.ID, advisoryIDs, nil)
 	database.CreateAdvisoryAccountData(t, system.RhAccountID, advisoryIDs, 1)
 	database.UpdateSystemAdvisoriesWhenPatched(t, int(system.ID), system.RhAccountID, advisoryIDs, &testDate)
 	// Update as-if the advisories had become patched
-	err := updateAdvisoryAccountData(database.Db, &system, advisoryIDs, []int{})
+	err := updateAdvisoryAccountData(database.Db, &system, advisoryIDs, []int64{})
 	assert.NoError(t, err)
 
 	database.CheckSystemAdvisoriesWhenPatched(t, system.ID, advisoryIDs, &testDate)
@@ -104,7 +104,7 @@ func TestUpdatePatchedSystemAdvisories(t *testing.T) {
 	database.UpdateSystemAdvisoriesWhenPatched(t, int(system.ID), system.RhAccountID, advisoryIDs, nil)
 
 	// Update as-if the advisories had become unpatched
-	err = updateAdvisoryAccountData(database.Db, &system, []int{}, advisoryIDs)
+	err = updateAdvisoryAccountData(database.Db, &system, []int64{}, advisoryIDs)
 	assert.NoError(t, err)
 
 	database.CheckAdvisoriesAccountData(t, system.RhAccountID, advisoryIDs, 1)
@@ -128,7 +128,7 @@ func TestEnsureSystemAdvisories(t *testing.T) {
 
 	rhAccountID := 1
 	systemID := int64(2)
-	advisoryIDs := []int{2, 3, 4}
+	advisoryIDs := []int64{2, 3, 4}
 	err := ensureSystemAdvisories(database.Db, rhAccountID, systemID, advisoryIDs)
 	assert.Nil(t, err)
 	database.CheckSystemAdvisoriesWhenPatched(t, systemID, advisoryIDs, nil)
